renaming ApiDocs to DocsApi so it can be reached through /api/test/Docs
